+++ src/app/app.routes.ts
@@ -1,8 +1,8 @@
 import { Routes } from '@angular/router';
-import { LoginPageComponent } from '@pages/login-page/login-page.component';
-import { MainPageComponent } from '@pages/main-page/main-page.component';
-import { RegistrationPageComponent } from '@pages/registration-page/registration-page.component';
-import { NotFoundPageComponent } from '@pages/not-found-page/not-found-page.component';
+import { LoginPageComponent } from '@pages/login-page';
+import { MainPageComponent } from '@pages/main-page';
+import { NotFoundPageComponent } from '@pages/not-found-page';
+import { RegistrationPageComponent } from '@pages/registration-page';
 import { canActivateAuth } from '@services/auth-service';
 
 export const routes: Routes = [
