+++ src/app/services/auth-service/access-guard/access.guard.ts
@@ -1,6 +1,6 @@
 import { inject } from '@angular/core';
-import { AuthService } from '../auth.service';
 import { Router, UrlTree } from '@angular/router';
+import { AuthService } from '@services/auth-service';
 
 export const canActivateAuth = (): boolean | UrlTree => {
   const isLoggedIn = inject(AuthService).isAuthorized();
